requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
python-dotenv==1.1.1
Pillow==11.0.0
gunicorn==23.0.0
whitenoise==6.7.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10